updated models,controllers,services to handle new fields of book, new Dto models for languages and tags

+++ backend/src/main/java/com/example/Luna/api/controller/BookController.java
@@ -28,7 +28,7 @@ public class BookController {
 
     // Build get book REST API
     @GetMapping("{id}")
-    public ResponseEntity<BookWithDiscountDto> getBookById(@PathVariable("id") int bookId) {
+    public ResponseEntity<BookWithDiscountDto> getBookById(@PathVariable("id") long bookId) {
         BookWithDiscountDto bookDto = bookService.getBookById(bookId);
         return ResponseEntity.ok(bookDto);
     }
