+++ backend/src/main/java/com/example/Luna/service/DiscountService.java
@@ -6,9 +6,9 @@ import com.example.Luna.api.model.Book;
 import java.util.List;
 
 public interface DiscountService {
-    DiscountDto getDiscountByBookId(int id);
+    DiscountDto getDiscountByBookId(long id);
     DiscountDto createDiscount(DiscountDto discountDto);
     List<DiscountDto> createDiscounts(List<DiscountDto> discountDtos);
-    DiscountDto createDiscountForBook(int bookId, DiscountDto discountDto);
+    DiscountDto createDiscountForBook(long bookId, DiscountDto discountDto);
     List<Book> getBooksOnActiveDiscount(List<Book> books);
 }
